msm: mdss: Fix logical NOT operator precedence in fb_event_callback
Resolved a warning caused by improper use of the logical NOT operator
in a comparison. The condition checking if 'ctrl_pdata->status_mode'
is not equal to 'ESD_REG_IRQ' was incorrectly structured, leading to
confusion and potential logical errors.

Updated the condition to ensure proper evaluation:
if (!(ctrl_pdata->status_mode == ESD_REG_IRQ))

This fixes the following warning:
../drivers/video/fbdev/msm/../../msm/mdss/mdss_dsi_status.c:182:8: warning: logical not is only applied to the left hand side of this comparison [-Wlogical-not-parentheses]

Signed-off-by: Robin <hanya8493@gmail.com>